Summary: Changed "displayedNumericalValue" of type "Double" to "Double?"
Task:
Change the computed instance variable displayValue to be an Optional Doublerather than a Double. Its value should be nil if the contents of display.text cannotbe interpreted as a Double. Setting its value to nil should clear the display out. You’llhave to modify the code that uses displayValue accordingly.

Did:
1) Changed "displayedNumericalValue" of type "Double" to "Double?"
2) added "if let displayedNumber = displayedNumericalValue {}" in "private func pressedOperation(sender: UIButton) {}" function to protect against cases where "displayedNumericalValue" is equal to nil, which happens when "displayLabel.text" does not contain a string that can be converted to a number
3) removed argument from "func clearDataAndResetCalculator(sender: UIButton) {}" to "func clearDataAndResetCalculator() {}" so that the function can be called when a button IS NOT pressed