Must be full of bugs -> Summary: "displayLabel.text" only display number up to 6 decimal places
Task:
Figure out from the documentation how to use the NSNumberFormatter class to formatyour display so that it only shows 6 digits after the decimal point (instead of showingall digits that can be represented in a Double). This will eliminate the need forAutoshrink in your display. While you’re at it, make it so that numbers that areintegers don’t have an unnecessary “.0” attached to them (e.g. show “4” rather than“4.0” as the result of the square root of sixteen). You can do all this for yourdescription in the CalculatorBrain as well.

Did:
1) used an instance of NSNumberFormatter to format "displayLabel.text" when converting "brain.result" of type Double to "displayLabel.text", which is of type String
2) added an extension to class "Double", that help round number to certain decimal places