Complete Assignment 2 (exclude extra credit)
Required Tasks

1. All of the changes to the Calculator made in lecture must be applied to your

Assignment 1 solution. This includes both the program var and properly setting access

control on all methods and properties. Get this fully functioning before proceeding to

the rest of the Required Tasks. And, as last week, type the changes in, do not copy/paste

from anywhere.

2. Do not change any non-private API in CalculatorBrain and continue to use the

Dictionary<String,Operation> as its primary internal data structure.

3. Your UI should always be in sync with your Model (the CalculatorBrain).

4. Add the capability to your CalculatorBrain to allow the input of variables. Do so by

implementing the following API in your CalculatorBrain ...

func setOperand(variableName: String)

var variableValues: Dictionary<String, Double>

These must do exactly what you would imagine they would: the first inputs a

“variable” as the operand (e.g. setOperand(“x”) would use a variable named x) and the

second lets users of the CalculatorBrain API set the value for any variable they wish

(e.g. brain.variableValues[“x”] = 35.0). Your CalculatorBrain must support any

number of variables. You can assume that no variable name will be the same as an

operation symbol.

5. The result var must now properly reflect variable values (from the variableValues

dictionary) whenever variables have been used (both in the past and in the future, see

8e below). If a variable has no value set in the dictionary, then use 0.0 as its value. If

the variableValues dictionary is changed, then the result will have to change to

reflect any new variable values.

6. Your description var must continue to work properly and should show the name of

the variable (not its value) wherever it appeared as input.

7. The program var added in lecture will need to be upgraded to support variables too.

8. Add two new buttons to your Calculator’s UI: →M and M. Don’t sacrifice any of the

required operation buttons from Assignment 1 to add these (though you may add

more operations buttons if you want). These two buttons will set and get (respectively)

a variable in the CalculatorBrain called M.

a. →M sets the value of the variable M in the brain to the current value of the display

and then shows the brain’s result in the display.

b. →M does not perform setOperand.

c. Touching M should setOperand(“M”) in the brain and then show the brain’s

result in the display.

d. →M and M are Controller mechanics, not Model mechanics (though they both use

the Model mechanic of variables).

e. This is not a very great “memory” button on our Calculator, but it’s good for

testing whether our variable function implemented above is working properly.

Examples ...

9 + M = √ ⇒ description is √(9+M), display is 3 because M is not set (and so is 0)

7 →M ⇒ display now shows 4 (the square root of 16), description is still √(9+M)

+ 14 = ⇒ display now shows 18, description is now √(9+M)+14

9. Make sure your C button from Assignment 1 works properly in this assignment. In

addition, it should remove any value for the M variable from the variableValues

Dictionary in the CalculatorBrain (not set it to zero or any other value). This will

allow you to test the case of an “unset” variable.

10.Add an Undo button to your Calculator. In Assignment 1’s Extra Credit, you might

have added a “backspace” button. Here we’re talking about combining both

backspace and actual undo into a single button. If the user is in the middle of

entering a number, this Undo button should be backspace. When the user is not in

the middle of entering a number, it should undo the last thing that was done in the

CalculatorBrain. Do not undo the storing of variable values (but DO undo the setting

of a variable as an operand).